+++ src/main/java/dev/sidgames/wmo/BlockLoader.java
@@ -14,6 +14,8 @@ public class BlockLoader {
         public static Block ALUMINUM_BLOCK;
         public static Block BAUXITE_ORE;
         public static Block BRONZE_BLOCK;
+        public static Block LEAD_BLOCK;
+        public static Block LEAD_ORE;
         public static Block LITHIUM_BLOCK;
         public static Block LITHIUM_ORE;
         public static Block PLATINUM_BLOCK;
@@ -34,6 +36,8 @@ public class BlockLoader {
                 ALUMINUM_BLOCK = Registry.register(Registry.BLOCK, new Identifier(Reference.MOD_ID, "aluminum_block"), new AluminumBlock());
                 BAUXITE_ORE = Registry.register(Registry.BLOCK, new Identifier(Reference.MOD_ID, "bauxite_ore"), new BauxiteOre());
                 BRONZE_BLOCK = Registry.register(Registry.BLOCK, new Identifier(Reference.MOD_ID, "bronze_block"), new BronzeBlock());
+                LEAD_BLOCK = Registry.register(Registry.BLOCK, new Identifier(Reference.MOD_ID, "lead_block"), new LeadBlock());
+                LEAD_ORE = Registry.register(Registry.BLOCK, new Identifier(Reference.MOD_ID, "lead_ore"), new LeadOre());
                 LITHIUM_BLOCK = Registry.register(Registry.BLOCK, new Identifier(Reference.MOD_ID, "lithium_block"), new LithiumBlock());
                 LITHIUM_ORE = Registry.register(Registry.BLOCK, new Identifier(Reference.MOD_ID, "lithium_ore"), new LithiumOre());
                 PLATINUM_BLOCK = Registry.register(Registry.BLOCK, new Identifier(Reference.MOD_ID, "platinum_block"), new PlatinumBlock());
@@ -57,6 +61,8 @@ public class BlockLoader {
                 Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "aluminum_block"), new BlockItem(ALUMINUM_BLOCK, new Item.Settings().group(ItemGroup.BUILDING_BLOCKS)));
                 Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "bauxite_ore"), new BlockItem(BAUXITE_ORE, new Item.Settings().group(ItemGroup.BUILDING_BLOCKS)));
                 Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "bronze_block"), new BlockItem(BRONZE_BLOCK, new Item.Settings().group(ItemGroup.BUILDING_BLOCKS)));
+                Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "lead_block"), new BlockItem(LEAD_BLOCK, new Item.Settings().group(ItemGroup.BUILDING_BLOCKS)));
+                Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "lead_ore"), new BlockItem(LEAD_ORE, new Item.Settings().group(ItemGroup.BUILDING_BLOCKS)));
                 Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "lithium_block"), new BlockItem(LITHIUM_BLOCK, new Item.Settings().group(ItemGroup.BUILDING_BLOCKS)));
                 Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "lithium_ore"), new BlockItem(LITHIUM_ORE, new Item.Settings().group(ItemGroup.BUILDING_BLOCKS)));
                 Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "platinum_block"), new BlockItem(PLATINUM_BLOCK, new Item.Settings().group(ItemGroup.BUILDING_BLOCKS)));