+++ src/main/java/dev/sidgames/wmo/ItemLoader.java
@@ -12,6 +12,8 @@ public class ItemLoader {
         Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "aluminum_ingot"), new AluminumIngot(new Item.Settings().group(ItemGroup.MATERIALS)));
         Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "aluminum_nugget"), new AluminumNugget(new Item.Settings().group(ItemGroup.MATERIALS)));
         Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "bronze_alloy"), new BronzeAlloy(new Item.Settings().group(ItemGroup.MATERIALS)));
+        Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "lead_ingot"), new LeadIngot(new Item.Settings().group(ItemGroup.MATERIALS)));
+        Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "lead_nugget"), new LeadNugget(new Item.Settings().group(ItemGroup.MATERIALS)));
         Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "lithium_ingot"), new LithiumIngot(new Item.Settings().group(ItemGroup.MATERIALS)));
         Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "lithium_nugget"), new LithiumNugget(new Item.Settings().group(ItemGroup.MATERIALS)));
         Registry.register(Registry.ITEM, new Identifier(Reference.MOD_ID, "platinum_ingot"), new PlatinumIngot(new Item.Settings().group(ItemGroup.MATERIALS)));