+++ src/main/java/dev/sidgames/wmo/block/LeadBlock.java
@@ -1,0 +1,11 @@
+package dev.sidgames.wmo.block;
+
+import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
+import net.minecraft.block.Material;
+
+public class LeadBlock extends BlockBase {
+    public LeadBlock() {
+        super(FabricBlockSettings.of(Material.METAL).hardness(5.0F).resistance(10.0F).requiresTool());
+    }
+
+}