+++ src/main/java/dev/sidgames/wmo/item/LeadIngot.java
@@ -1,0 +1,8 @@
+package dev.sidgames.wmo.item;
+
+public class LeadIngot extends ItemBase {
+    public LeadIngot(Settings settings) {
+        super(settings);
+    }
+}
+