+++ src/main/java/dev/sidgames/wmo/item/LeadNugget.java
@@ -1,0 +1,8 @@
+package dev.sidgames.wmo.item;
+
+public class LeadNugget extends ItemBase {
+    public LeadNugget(Settings settings) {
+        super(settings);
+    }
+
+}